Estimate prior genotype probabilities by maximum likelihood

We maximize the marginal likelihood to obtain estimates of the prior probabilities of each genotype at each locus. The general categorical class uses the method of Li (2011). The proportional normal class uses the method of Gerard and Ferrão (2020).

est_prior_mat(gl, method = c("pnorm", "general"), log = TRUE)

Arguments

gl

A three dimensional array of genotype log-likelihoods. Element gl[i, j, k] is the genotype log-likelihood of dosage k for individual j at SNP i.

method

Should we assume the class of general categorical prior distributions ("general") or the class of proportional normal distributions ("pnorm")?

log

A logical. Should we log the results or not?

Value

A matrix of prior genotype probabilities. Element (i, k)

is the prior probability of genotype k at locus i.

References

Gerard, D., & Ferrão, L. F. V. (2020). Priors for genotyping polyploids. Bioinformatics, 36(6), 1795-1800. doi:10.1093/bioinformatics/btz852

Li, H. (2011). A statistical framework for SNP calling, mutation discovery, association mapping and population genetical parameter estimation from sequencing data. Bioinformatics, 27(21), 2987-2993. doi:10.1093/bioinformatics/btr509

See also

Author

David Gerard and D. Thomas Scartz

Examples

data("glike")
est_prior_mat(glike)
#>             [,1]       [,2]       [,3]       [,4]        [,5]
#>  [1,]  -3.213015  -1.616097 -0.9568160 -1.2351711  -2.4511626
#>  [2,]  -3.213015  -1.616097 -0.9568160 -1.2351711  -2.4511626
#>  [3,]  -3.813907  -2.369151 -1.4508240 -1.0589247  -1.1934534
#>  [4,]  -2.611388  -1.189717 -0.8852265 -1.6979167  -3.6277874
#>  [5,]  -4.547391  -2.863992 -1.6980130 -1.0494556  -0.9183193
#>  [6,]  -4.547391  -2.863992 -1.6980130 -1.0494556  -0.9183193
#>  [7,]  -3.747561  -2.338272 -1.4446214 -1.0666088  -1.2042344
#>  [8,]  -3.747561  -2.338272 -1.4446214 -1.0666088  -1.2042344
#>  [9,] -43.257699 -14.321332 -0.3273840 -1.2758538 -17.1667416
#> [10,]  -5.251005  -3.082684 -1.6499255 -0.9527288  -0.9910942