Simulate genotype likelihood array for individuals given a banded correlation structure.

Simulates genotypes using geno_sim(), then simulates sequencing data using the model of Gerard et al (2018), then fits flexdog() on these data to obtain genotype likelihoods.

glsim(
  rho,
  alpha,
  nloc,
  ploidy,
  nind,
  rdepth = 10,
  seq = 0.001,
  bias = 1,
  od = 0.001
)

Arguments

rho

The correlation between adjacent loci. If loci are a positions apart, then the correlation between them is rho^a.

alpha

The allele frequency at each locus.

nloc

The number of loci to simulate.

ploidy

The ploidy of the individual.

nind

The number of individuals to sample.

rdepth

The read-depth of the sequencing reads used to create the genoype likelihoods. Lower means greater uncertainty.

seq

The sequencing error rate.

bias

The allele bias.

od

The overdispersion rate.

References

Gerard, D., Ferrão, L. F. V., Garcia, A. A. F., & Stephens, M. (2018). Genotyping polyploids from messy sequencing data. Genetics, 210(3), 789-807. doi:genetics.118.301468

Author

David Gerard and Hanwei Hu

Examples

glsim(rho = 0.9, alpha = 0.5, nloc = 10, ploidy = 4, nind = 25)
#> Loading required package: foreach
#> Loading required package: future
#> Loading required package: rngtools
#> , , variable = logL_0
#> 
#>     ind
#> snp          1         10         11         12            13        14
#>   1  -40.89139  -9.340121 -24.259038 -14.173757 -14.173757045 -24.25904
#>   2  -40.89139 -14.173757  -9.340121 -24.259038  -0.009960197 -19.13985
#>   3  -40.89139 -35.082125  -9.340121  -4.623107  -0.009960197 -35.08212
#>   4  -40.89139  -9.340121  -4.623107  -9.340121  -0.009960197 -35.08212
#>   5  -35.08212 -24.259038 -24.259038  -9.340121  -0.009960197 -29.55952
#>   6  -29.55952 -29.559521 -19.139845 -40.891394  -0.009960197 -29.55952
#>   7  -47.10511 -29.559521 -19.139845 -24.259038  -9.340121383 -19.13985
#>   8  -24.25904 -24.259038 -24.259038 -19.139845 -14.173757045 -14.17376
#>   9  -40.89139 -24.259038 -19.139845 -35.082125 -14.173757045 -24.25904
#>   10 -24.25904 -24.259038 -29.559521  -9.340121 -24.259037560 -35.08212
#>     ind
#> snp            15         16         17            18         19          2
#>   1  -0.009960197  -4.623107 -19.139845 -19.139845224  -4.623107  -9.340121
#>   2  -0.009960197  -9.340121 -14.173757  -9.340121383 -24.259038  -9.340121
#>   3  -0.009960197 -14.173757 -29.559521 -24.259037560 -35.082125 -14.173757
#>   4  -0.009960197 -14.173757 -35.082125 -35.082124892 -19.139845 -19.139845
#>   5  -0.009960197 -24.259038 -19.139845 -29.559520607 -24.259038 -29.559521
#>   6  -0.009960197 -19.139845 -24.259038 -29.559520607 -19.139845 -29.559521
#>   7  -0.009960197  -4.623107 -35.082125 -35.082124892 -24.259038 -35.082125
#>   8  -0.009960197 -24.259038 -19.139845 -19.139845224 -29.559521 -40.891394
#>   9  -0.009960197  -9.340121 -24.259038  -0.009960197 -24.259038 -24.259038
#>   10 -0.009960197 -19.139845  -9.340121 -29.559520607 -29.559521 -35.082125
#>     ind
#> snp             20            21        22        23        24        25
#>   1  -24.259037560 -40.891393890 -19.13985 -35.08212 -24.25904 -35.08212
#>   2  -29.559520607 -40.891393890 -24.25904 -29.55952 -24.25904 -19.13985
#>   3  -29.559520607  -0.009960197 -24.25904 -40.89139 -24.25904 -24.25904
#>   4   -9.340121383 -19.139845224 -14.17376 -35.08212 -40.89139 -54.01097
#>   5  -14.173757045 -35.082124892 -35.08212 -40.89139 -40.89139 -54.01097
#>   6  -14.173757045 -14.173757045 -24.25904 -40.89139 -54.01097 -54.01097
#>   7   -4.623107490 -19.139845224 -14.17376 -54.01097 -35.08212 -54.01097
#>   8   -0.009960197 -40.891393890 -19.13985 -47.10511 -47.10511 -54.01097
#>   9  -19.139845224 -29.559520607 -35.08212 -19.13985 -54.01097 -54.01097
#>   10 -19.139845224 -35.082124892 -29.55952 -19.13985 -54.01097 -54.01097
#>     ind
#> snp           3          4         5          6          7          8         9
#>   1   -4.623107 -35.082125 -24.25904  -4.623107 -14.173757 -40.891394 -40.89139
#>   2   -9.340121 -29.559521 -40.89139 -29.559521 -14.173757  -9.340121 -47.10511
#>   3   -4.623107 -24.259038 -40.89139 -29.559521  -9.340121 -24.259038 -54.01097
#>   4   -4.623107 -24.259038 -40.89139 -14.173757 -29.559521 -14.173757 -35.08212
#>   5  -14.173757 -19.139845 -24.25904 -29.559521 -14.173757  -9.340121 -40.89139
#>   6   -4.623107 -35.082125 -35.08212  -9.340121 -24.259038  -4.623107 -47.10511
#>   7  -19.139845 -19.139845 -47.10511 -14.173757 -35.082125 -19.139845 -35.08212
#>   8   -4.623107  -9.340121 -54.01097 -24.259038 -35.082125  -4.623107 -35.08212
#>   9  -19.139845 -40.891394 -24.25904 -29.559521  -9.340121  -4.623107 -40.89139
#>   10 -19.139845 -24.259038 -19.13985 -24.259038 -14.173757 -24.259038 -54.01097
#> 
#> , , variable = logL_1
#> 
#>     ind
#> snp           1        10        11        12        13        14        15
#>   1   -7.777188 -1.272365 -2.825649 -1.388829 -1.388829 -2.825649 -2.868545
#>   2   -7.777188 -1.388829 -1.272365 -2.825649 -2.868545 -1.921225 -2.868545
#>   3   -7.777188 -5.725331 -1.272365 -1.673856 -2.868545 -5.725331 -2.868545
#>   4   -7.777188 -1.272365 -1.673856 -1.272365 -2.868545 -5.725331 -2.868545
#>   5   -5.725331 -2.825649 -2.825649 -1.272365 -2.868545 -4.089463 -2.868545
#>   6   -4.089463 -4.089463 -1.921225 -7.777188 -2.868545 -4.089463 -2.868545
#>   7  -10.347079 -4.089463 -1.921225 -2.825649 -1.272365 -1.921225 -2.868545
#>   8   -2.825649 -2.825649 -2.825649 -1.921225 -1.388829 -1.388829 -2.868545
#>   9   -7.777188 -2.825649 -1.921225 -5.725331 -1.388829 -2.825649 -2.868545
#>   10  -2.825649 -2.825649 -4.089463 -1.272365 -2.825649 -5.725331 -2.868545
#>     ind
#> snp         16        17        18        19         2        20        21
#>   1  -1.673856 -1.921225 -1.921225 -1.673856 -1.272365 -2.825649 -7.777188
#>   2  -1.272365 -1.388829 -1.272365 -2.825649 -1.272365 -4.089463 -7.777188
#>   3  -1.388829 -4.089463 -2.825649 -5.725331 -1.388829 -4.089463 -2.868545
#>   4  -1.388829 -5.725331 -5.725331 -1.921225 -1.921225 -1.272365 -1.921225
#>   5  -2.825649 -1.921225 -4.089463 -2.825649 -4.089463 -1.388829 -5.725331
#>   6  -1.921225 -2.825649 -4.089463 -1.921225 -4.089463 -1.388829 -1.388829
#>   7  -1.673856 -5.725331 -5.725331 -2.825649 -5.725331 -1.673856 -1.921225
#>   8  -2.825649 -1.921225 -1.921225 -4.089463 -7.777188 -2.868545 -7.777188
#>   9  -1.272365 -2.825649 -2.868545 -2.825649 -2.825649 -1.921225 -4.089463
#>   10 -1.921225 -1.272365 -4.089463 -4.089463 -5.725331 -1.921225 -5.725331
#>     ind
#> snp         22         23         24         25         3         4          5
#>   1  -1.921225  -5.725331  -2.825649  -5.725331 -1.673856 -5.725331  -2.825649
#>   2  -2.825649  -4.089463  -2.825649  -1.921225 -1.272365 -4.089463  -7.777188
#>   3  -2.825649  -7.777188  -2.825649  -2.825649 -1.673856 -2.825649  -7.777188
#>   4  -1.388829  -5.725331  -7.777188 -13.710279 -1.673856 -2.825649  -7.777188
#>   5  -5.725331  -7.777188  -7.777188 -13.710279 -1.388829 -1.921225  -2.825649
#>   6  -2.825649  -7.777188 -13.710279 -13.710279 -1.673856 -5.725331  -5.725331
#>   7  -1.388829 -13.710279  -5.725331 -13.710279 -1.921225 -1.921225 -10.347079
#>   8  -1.921225 -10.347079 -10.347079 -13.710279 -1.673856 -1.272365 -13.710279
#>   9  -5.725331  -1.921225 -13.710279 -13.710279 -1.921225 -7.777188  -2.825649
#>   10 -4.089463  -1.921225 -13.710279 -13.710279 -1.921225 -2.825649  -1.921225
#>     ind
#> snp          6         7         8          9
#>   1  -1.673856 -1.388829 -7.777188  -7.777188
#>   2  -4.089463 -1.388829 -1.272365 -10.347079
#>   3  -4.089463 -1.272365 -2.825649 -13.710279
#>   4  -1.388829 -4.089463 -1.388829  -5.725331
#>   5  -4.089463 -1.388829 -1.272365  -7.777188
#>   6  -1.272365 -2.825649 -1.673856 -10.347079
#>   7  -1.388829 -5.725331 -1.921225  -5.725331
#>   8  -2.825649 -5.725331 -1.673856  -5.725331
#>   9  -4.089463 -1.272365 -1.673856  -7.777188
#>   10 -2.825649 -1.388829 -2.825649 -13.710279
#> 
#> , , variable = logL_2
#> 
#>     ind
#> snp          1        10        11        12        13        14       15
#>   1  -3.111935 -3.111935 -1.407025 -2.141026 -2.141026 -1.407025 -6.88685
#>   2  -3.111935 -2.141026 -3.111935 -1.407025 -6.886850 -1.587363 -6.88685
#>   3  -3.111935 -2.141026 -3.111935 -4.602123 -6.886850 -2.141026 -6.88685
#>   4  -3.111935 -3.111935 -4.602123 -3.111935 -6.886850 -2.141026 -6.88685
#>   5  -2.141026 -1.407025 -1.407025 -3.111935 -6.886850 -1.587363 -6.88685
#>   6  -1.587363 -1.587363 -1.587363 -3.111935 -6.886850 -1.587363 -6.88685
#>   7  -4.602123 -1.587363 -1.587363 -1.407025 -3.111935 -1.587363 -6.88685
#>   8  -1.407025 -1.407025 -1.407025 -1.587363 -2.141026 -2.141026 -6.88685
#>   9  -3.111935 -1.407025 -1.587363 -2.141026 -2.141026 -1.407025 -6.88685
#>   10 -1.407025 -1.407025 -1.587363 -3.111935 -1.407025 -2.141026 -6.88685
#>     ind
#> snp         16        17        18        19         2        20        21
#>   1  -4.602123 -1.587363 -1.587363 -4.602123 -3.111935 -1.407025 -3.111935
#>   2  -3.111935 -2.141026 -3.111935 -1.407025 -3.111935 -1.587363 -3.111935
#>   3  -2.141026 -1.587363 -1.407025 -2.141026 -2.141026 -1.587363 -6.886850
#>   4  -2.141026 -2.141026 -2.141026 -1.587363 -1.587363 -3.111935 -1.587363
#>   5  -1.407025 -1.587363 -1.587363 -1.407025 -1.587363 -2.141026 -2.141026
#>   6  -1.587363 -1.407025 -1.587363 -1.587363 -1.587363 -2.141026 -2.141026
#>   7  -4.602123 -2.141026 -2.141026 -1.407025 -2.141026 -4.602123 -1.587363
#>   8  -1.407025 -1.587363 -1.587363 -1.587363 -3.111935 -6.886850 -3.111935
#>   9  -3.111935 -1.407025 -6.886850 -1.407025 -1.407025 -1.587363 -1.587363
#>   10 -1.587363 -3.111935 -1.587363 -1.587363 -2.141026 -1.587363 -2.141026
#>     ind
#> snp         22        23        24        25         3         4         5
#>   1  -1.587363 -2.141026 -1.407025 -2.141026 -4.602123 -2.141026 -1.407025
#>   2  -1.407025 -1.587363 -1.407025 -1.587363 -3.111935 -1.587363 -3.111935
#>   3  -1.407025 -3.111935 -1.407025 -1.407025 -4.602123 -1.407025 -3.111935
#>   4  -2.141026 -2.141026 -3.111935 -6.886850 -4.602123 -1.407025 -3.111935
#>   5  -2.141026 -3.111935 -3.111935 -6.886850 -2.141026 -1.587363 -1.407025
#>   6  -1.407025 -3.111935 -6.886850 -6.886850 -4.602123 -2.141026 -2.141026
#>   7  -2.141026 -6.886850 -2.141026 -6.886850 -1.587363 -1.587363 -4.602123
#>   8  -1.587363 -4.602123 -4.602123 -6.886850 -4.602123 -3.111935 -6.886850
#>   9  -2.141026 -1.587363 -6.886850 -6.886850 -1.587363 -3.111935 -1.407025
#>   10 -1.587363 -1.587363 -6.886850 -6.886850 -1.587363 -1.407025 -1.587363
#>     ind
#> snp          6         7         8         9
#>   1  -4.602123 -2.141026 -3.111935 -3.111935
#>   2  -1.587363 -2.141026 -3.111935 -4.602123
#>   3  -1.587363 -3.111935 -1.407025 -6.886850
#>   4  -2.141026 -1.587363 -2.141026 -2.141026
#>   5  -1.587363 -2.141026 -3.111935 -3.111935
#>   6  -3.111935 -1.407025 -4.602123 -4.602123
#>   7  -2.141026 -2.141026 -1.587363 -2.141026
#>   8  -1.407025 -2.141026 -4.602123 -2.141026
#>   9  -1.587363 -3.111935 -4.602123 -3.111935
#>   10 -1.407025 -2.141026 -1.407025 -6.886850
#> 
#> , , variable = logL_3
#> 
#>     ind
#> snp          1        10         11         12         13        14        15
#>   1  -1.272365 -7.777188  -2.825649  -5.725331  -5.725331 -2.825649 -13.71028
#>   2  -1.272365 -5.725331  -7.777188  -2.825649 -13.710279 -4.089463 -13.71028
#>   3  -1.272365 -1.388829  -7.777188 -10.347079 -13.710279 -1.388829 -13.71028
#>   4  -1.272365 -7.777188 -10.347079  -7.777188 -13.710279 -1.388829 -13.71028
#>   5  -1.388829 -2.825649  -2.825649  -7.777188 -13.710279 -1.921225 -13.71028
#>   6  -1.921225 -1.921225  -4.089463  -1.272365 -13.710279 -1.921225 -13.71028
#>   7  -1.673856 -1.921225  -4.089463  -2.825649  -7.777188 -4.089463 -13.71028
#>   8  -2.825649 -2.825649  -2.825649  -4.089463  -5.725331 -5.725331 -13.71028
#>   9  -1.272365 -2.825649  -4.089463  -1.388829  -5.725331 -2.825649 -13.71028
#>   10 -2.825649 -2.825649  -1.921225  -7.777188  -2.825649 -1.388829 -13.71028
#>     ind
#> snp          16        17         18         19         2         20         21
#>   1  -10.347079 -4.089463  -4.089463 -10.347079 -7.777188  -2.825649  -1.272365
#>   2   -7.777188 -5.725331  -7.777188  -2.825649 -7.777188  -1.921225  -1.272365
#>   3   -5.725331 -1.921225  -2.825649  -1.388829 -5.725331  -1.921225 -13.710279
#>   4   -5.725331 -1.388829  -1.388829  -4.089463 -4.089463  -7.777188  -4.089463
#>   5   -2.825649 -4.089463  -1.921225  -2.825649 -1.921225  -5.725331  -1.388829
#>   6   -4.089463 -2.825649  -1.921225  -4.089463 -1.921225  -5.725331  -5.725331
#>   7  -10.347079 -1.388829  -1.388829  -2.825649 -1.388829 -10.347079  -4.089463
#>   8   -2.825649 -4.089463  -4.089463  -1.921225 -1.272365 -13.710279  -1.272365
#>   9   -7.777188 -2.825649 -13.710279  -2.825649 -2.825649  -4.089463  -1.921225
#>   10  -4.089463 -7.777188  -1.921225  -1.921225 -1.388829  -4.089463  -1.388829
#>     ind
#> snp         22        23        24        25          3         4         5
#>   1  -4.089463 -1.388829 -2.825649 -1.388829 -10.347079 -1.388829 -2.825649
#>   2  -2.825649 -1.921225 -2.825649 -4.089463  -7.777188 -1.921225 -1.272365
#>   3  -2.825649 -1.272365 -2.825649 -2.825649 -10.347079 -2.825649 -1.272365
#>   4  -5.725331 -1.388829 -1.272365 -2.868545 -10.347079 -2.825649 -1.272365
#>   5  -1.388829 -1.272365 -1.272365 -2.868545  -5.725331 -4.089463 -2.825649
#>   6  -2.825649 -1.272365 -2.868545 -2.868545 -10.347079 -1.388829 -1.388829
#>   7  -5.725331 -2.868545 -1.388829 -2.868545  -4.089463 -4.089463 -1.673856
#>   8  -4.089463 -1.673856 -1.673856 -2.868545 -10.347079 -7.777188 -2.868545
#>   9  -1.388829 -4.089463 -2.868545 -2.868545  -4.089463 -1.272365 -2.825649
#>   10 -1.921225 -4.089463 -2.868545 -2.868545  -4.089463 -2.825649 -4.089463
#>     ind
#> snp           6         7          8         9
#>   1  -10.347079 -5.725331  -1.272365 -1.272365
#>   2   -1.921225 -5.725331  -7.777188 -1.673856
#>   3   -1.921225 -7.777188  -2.825649 -2.868545
#>   4   -5.725331 -1.921225  -5.725331 -1.388829
#>   5   -1.921225 -5.725331  -7.777188 -1.272365
#>   6   -7.777188 -2.825649 -10.347079 -1.673856
#>   7   -5.725331 -1.388829  -4.089463 -1.388829
#>   8   -2.825649 -1.388829 -10.347079 -1.388829
#>   9   -1.921225 -7.777188 -10.347079 -1.272365
#>   10  -2.825649 -5.725331  -2.825649 -2.868545
#> 
#> , , variable = logL_4
#> 
#>     ind
#> snp           1        10        11         12        13        14        15
#>   1   -9.340121 -40.89139 -24.25904 -35.082125 -35.08212 -24.25904 -54.01097
#>   2   -9.340121 -35.08212 -40.89139 -24.259038 -54.01097 -29.55952 -54.01097
#>   3   -9.340121 -14.17376 -40.89139 -47.105114 -54.01097 -14.17376 -54.01097
#>   4   -9.340121 -40.89139 -47.10511 -40.891394 -54.01097 -14.17376 -54.01097
#>   5  -14.173757 -24.25904 -24.25904 -40.891394 -54.01097 -19.13985 -54.01097
#>   6  -19.139845 -19.13985 -29.55952  -9.340121 -54.01097 -19.13985 -54.01097
#>   7   -4.623107 -19.13985 -29.55952 -24.259038 -40.89139 -29.55952 -54.01097
#>   8  -24.259038 -24.25904 -24.25904 -29.559521 -35.08212 -35.08212 -54.01097
#>   9   -9.340121 -24.25904 -29.55952 -14.173757 -35.08212 -24.25904 -54.01097
#>   10 -24.259038 -24.25904 -19.13985 -40.891394 -24.25904 -14.17376 -54.01097
#>     ind
#> snp         16        17        18        19          2        20         21
#>   1  -47.10511 -29.55952 -29.55952 -47.10511 -40.891394 -24.25904  -9.340121
#>   2  -40.89139 -35.08212 -40.89139 -24.25904 -40.891394 -19.13985  -9.340121
#>   3  -35.08212 -19.13985 -24.25904 -14.17376 -35.082125 -19.13985 -54.010968
#>   4  -35.08212 -14.17376 -14.17376 -29.55952 -29.559521 -40.89139 -29.559521
#>   5  -24.25904 -29.55952 -19.13985 -24.25904 -19.139845 -35.08212 -14.173757
#>   6  -29.55952 -24.25904 -19.13985 -29.55952 -19.139845 -35.08212 -35.082125
#>   7  -47.10511 -14.17376 -14.17376 -24.25904 -14.173757 -47.10511 -29.559521
#>   8  -24.25904 -29.55952 -29.55952 -19.13985  -9.340121 -54.01097  -9.340121
#>   9  -40.89139 -24.25904 -54.01097 -24.25904 -24.259038 -29.55952 -19.139845
#>   10 -29.55952 -40.89139 -19.13985 -19.13985 -14.173757 -29.55952 -14.173757
#>     ind
#> snp         22            23            24            25         3          4
#>   1  -29.55952 -14.173757045 -24.259037560 -14.173757045 -47.10511 -14.173757
#>   2  -24.25904 -19.139845224 -24.259037560 -29.559520607 -40.89139 -19.139845
#>   3  -24.25904  -9.340121383 -24.259037560 -24.259037560 -47.10511 -24.259038
#>   4  -35.08212 -14.173757045  -9.340121383  -0.009960197 -47.10511 -24.259038
#>   5  -14.17376  -9.340121383  -9.340121383  -0.009960197 -35.08212 -29.559521
#>   6  -24.25904  -9.340121383  -0.009960197  -0.009960197 -47.10511 -14.173757
#>   7  -35.08212  -0.009960197 -14.173757045  -0.009960197 -29.55952 -29.559521
#>   8  -29.55952  -4.623107490  -4.623107490  -0.009960197 -47.10511 -40.891394
#>   9  -14.17376 -29.559520607  -0.009960197  -0.009960197 -29.55952  -9.340121
#>   10 -19.13985 -29.559520607  -0.009960197  -0.009960197 -29.55952 -24.259038
#>     ind
#> snp              5         6         7          8             9
#>   1  -24.259037560 -47.10511 -35.08212  -9.340121  -9.340121383
#>   2   -9.340121383 -19.13985 -35.08212 -40.891394  -4.623107490
#>   3   -9.340121383 -19.13985 -40.89139 -24.259038  -0.009960197
#>   4   -9.340121383 -35.08212 -19.13985 -35.082125 -14.173757045
#>   5  -24.259037560 -19.13985 -35.08212 -40.891394  -9.340121383
#>   6  -14.173757045 -40.89139 -24.25904 -47.105114  -4.623107490
#>   7   -4.623107490 -35.08212 -14.17376 -29.559521 -14.173757045
#>   8   -0.009960197 -24.25904 -14.17376 -47.105114 -14.173757045
#>   9  -24.259037560 -19.13985 -40.89139 -47.105114  -9.340121383
#>   10 -29.559520607 -24.25904 -35.08212 -24.259038  -0.009960197
#>