A function to simulate genotype log-likelihoods under a given correlation under HWE for two loci

A user inputs the correlation (rho), the number of individuals (nind), the ploidy (ploidy), the allele frequencies at two loci (pA and pB), and the "read-depth" (larger corresponds to being more sure of the genotypes) (rdepth). Then this function will simulate some genotype log-likelihoods.

glsim_pairwise(
  rho,
  nind,
  ploidy,
  pA,
  pB,
  rdepth = 10,
  seq = 0.001,
  bias = 1,
  od = 0.001
)

Arguments

rho

The true correlation.

nind

The number of individuals to simulate.

ploidy

The ploidy of the species.

pA

The allele frequency of the first locus.

pB

The allele frequency of the second locus.

rdepth

The read depth to simulate (larger means more certain).

seq

The sequencing error rate.

bias

The allele bias.

od

The overdispersion parameter.

Value

A list with some or all of the following elements

lA

The genotype log-likelihoods at the first locus. lA[i, k] is the genotype log-likelihood for individual i at genotype k - 1.

lB

The genotype log-likelihoods at the second locus. lB[i, K] is the genotype log-likelihood for individual i at genotype k - 1.

rho

The true Pearson correlation that the user provided.

delta

The corresponding LD coefficient.

gA

The vector of true genotypes for locus A

gB

The vector of true genotypes for locus B

pA

The matrix of posterior genotype probabilities for locus A

pB

The matrix of posterior genotype probabilities for locus B

Author

David Gerard

Examples

glout <- glsim_pairwise(rho = 0.9, nind = 10000, ploidy = 6, pA = 0.5, pB = 0.5)

cor(glout$gA, glout$gB)
#> [1] 0.9005799