Return the true correlation from hap_sim().

The correlation is rho^a, where a is the number of positions away two SNPs are.

hap_truecor(rho, nloc)

Arguments

rho

The correlation between adjacent loci. If loci are a positions apart, then the correlation between them is rho^a.

nloc

The number of loci to simulate.

Value

A matrix of size nloc by nloc that contains the true correlation of the data simulated in hap_sim().

Author

David Gerard and Hanwei Hu

Examples

hap_truecor(0.9, 10)
#>            [,1]      [,2]      [,3]     [,4]    [,5]    [,6]     [,7]      [,8]
#>  [1,] 1.0000000 0.9000000 0.8100000 0.729000 0.65610 0.59049 0.531441 0.4782969
#>  [2,] 0.9000000 1.0000000 0.9000000 0.810000 0.72900 0.65610 0.590490 0.5314410
#>  [3,] 0.8100000 0.9000000 1.0000000 0.900000 0.81000 0.72900 0.656100 0.5904900
#>  [4,] 0.7290000 0.8100000 0.9000000 1.000000 0.90000 0.81000 0.729000 0.6561000
#>  [5,] 0.6561000 0.7290000 0.8100000 0.900000 1.00000 0.90000 0.810000 0.7290000
#>  [6,] 0.5904900 0.6561000 0.7290000 0.810000 0.90000 1.00000 0.900000 0.8100000
#>  [7,] 0.5314410 0.5904900 0.6561000 0.729000 0.81000 0.90000 1.000000 0.9000000
#>  [8,] 0.4782969 0.5314410 0.5904900 0.656100 0.72900 0.81000 0.900000 1.0000000
#>  [9,] 0.4304672 0.4782969 0.5314410 0.590490 0.65610 0.72900 0.810000 0.9000000
#> [10,] 0.3874205 0.4304672 0.4782969 0.531441 0.59049 0.65610 0.729000 0.8100000
#>            [,9]     [,10]
#>  [1,] 0.4304672 0.3874205
#>  [2,] 0.4782969 0.4304672
#>  [3,] 0.5314410 0.4782969
#>  [4,] 0.5904900 0.5314410
#>  [5,] 0.6561000 0.5904900
#>  [6,] 0.7290000 0.6561000
#>  [7,] 0.8100000 0.7290000
#>  [8,] 0.9000000 0.8100000
#>  [9,] 1.0000000 0.9000000
#> [10,] 0.9000000 1.0000000