Naive approach to estimate LD using posterior moments.

Here, we just use posterior means as a proxy for the genotypes, and calculate these LD estimates as if the genotypes were known. This is not a good way to do things if there is any genotype uncertainty (which is almost always the case in polyploids).

ldfast_naive(gp, type = c("r", "r2", "z", "D", "Dprime"))

Arguments

gp

A three-way array with dimensions SNPs by individuals by dosage. That is, gp[i, j, k] is the posterior probability of dosage k-1 for individual j at SNP i.

type

What LD measure should we estimate?

"r"

The Pearson correlation.

"r2"

The squared Pearson correlation.

"z"

The Fisher-z transformed Pearson correlation.

"D"

The LD coefficient.

"Dprime"

The standardized LD coefficient.

Note that these are all composite measures of LD (see the description in ldest()).

Value

A list with the following elements.

ldmat

The bias-corrected LD matrix.

Author

David Gerard

Examples

data("glike")
gp <- gl_to_gp(gl = glike, prior_mat = "estimate_pnorm")
ldfast_naive(gp = gp)
#> $ldmat
#>             [,1]       [,2]       [,3]         [,4]      [,5]      [,6]
#>  [1,] 1.00000000 1.00000000 0.42862407  0.823237426 0.5983991 0.5983991
#>  [2,] 1.00000000 1.00000000 0.42862407  0.823237426 0.5983991 0.5983991
#>  [3,] 0.42862407 0.42862407 1.00000000  0.569075315 0.8879786 0.8879786
#>  [4,] 0.82323743 0.82323743 0.56907532  1.000000000 0.5454147 0.5454147
#>  [5,] 0.59839914 0.59839914 0.88797855  0.545414700 1.0000000 1.0000000
#>  [6,] 0.59839914 0.59839914 0.88797855  0.545414700 1.0000000 1.0000000
#>  [7,] 0.42404788 0.42404788 0.97149161  0.576466698 0.8835920 0.8835920
#>  [8,] 0.42404788 0.42404788 0.97149161  0.576466698 0.8835920 0.8835920
#>  [9,] 0.09693651 0.09693651 0.08726337 -0.001911192 0.1260134 0.1260134
#> [10,] 0.53656274 0.53656274 0.79518829  0.521588524 0.9288950 0.9288950
#>             [,7]       [,8]         [,9]     [,10]
#>  [1,] 0.42404788 0.42404788  0.096936509 0.5365627
#>  [2,] 0.42404788 0.42404788  0.096936509 0.5365627
#>  [3,] 0.97149161 0.97149161  0.087263374 0.7951883
#>  [4,] 0.57646670 0.57646670 -0.001911192 0.5215885
#>  [5,] 0.88359197 0.88359197  0.126013372 0.9288950
#>  [6,] 0.88359197 0.88359197  0.126013372 0.9288950
#>  [7,] 1.00000000 1.00000000  0.074136925 0.8212752
#>  [8,] 1.00000000 1.00000000  0.074136925 0.8212752
#>  [9,] 0.07413693 0.07413693  1.000000000 0.0930604
#> [10,] 0.82127524 0.82127524  0.093060400 1.0000000
#>