Scalable LD Calculations Assuming a Uniform Prior

Here, we assume that gp was calculated assuming a discrete uniform prior over the genotypes. This is different from ldfast() where we assumed an adaptive prior had been used.

ldfast_unif(
  gp,
  type = c("r", "r2", "z", "D", "Dprime"),
  shrinkrr = TRUE,
  se = TRUE,
  thresh = TRUE,
  upper = 10,
  mode = c("zero", "estimate")
)

Arguments

gp

A three-way array with dimensions SNPs by individuals by dosage. That is, gp[i, j, k] is the posterior probability of dosage k-1 for individual j at SNP i.

type

What LD measure should we estimate?

"r"

The Pearson correlation.

"r2"

The squared Pearson correlation.

"z"

The Fisher-z transformed Pearson correlation.

"D"

The LD coefficient.

"Dprime"

The standardized LD coefficient.

Note that these are all composite measures of LD (see the description in ldest()).

shrinkrr

A logical. Should we use adaptive shrinkage (Stephens, 2016) to shrink the reliability ratios (TRUE) or keep the raw reliability ratios (FALSE). Defaults to TRUE.

se

Should we also return a matrix of standard errors (TRUE) or not (FALSE)? It is faster to not return standard errors. Defaults to TRUE.

thresh

A logical. Should we apply an upper bound on the reliability ratios (TRUE) or not (FALSE).

upper

The upper bound on the reliability ratios if thresh = TRUE. The default is a generous 10.

mode

A character. Only applies if shrinkrr = TRUE. When using hierarchical shrinkage on the log of the reliability ratios, should we use zero as the mode (mode = "zero") or estimate it using the procedure of Robertson and Cryer (1974) (mode = "estimate")?

Value

A list with some or all of the following elements:

ldmat

The bias-corrected LD matrix.

rr

The estimated reliability ratio for each SNP. This is the multiplicative factor applied to the naive LD estimate for each SNP.

See also

Author

David Gerard and Hanwei Hu

Examples

data("glike")
gp <- gl_to_gp(gl = glike)
ldfast_unif(gp = gp, se = FALSE)
#> $ldmat
#>            [,1]      [,2]       [,3]         [,4]      [,5]      [,6]
#>  [1,] 1.0000000 1.0000000 0.43383606  0.832466699 0.6031796 0.6031796
#>  [2,] 1.0000000 1.0000000 0.43383606  0.832466699 0.6031796 0.6031796
#>  [3,] 0.4338361 0.4338361 1.00000000  0.570536135 0.8931431 0.8931431
#>  [4,] 0.8324667 0.8324667 0.57053613  1.000000000 0.5444601 0.5444601
#>  [5,] 0.6031796 0.6031796 0.89314309  0.544460102 1.0000000 1.0000000
#>  [6,] 0.6031796 0.6031796 0.89314309  0.544460102 1.0000000 1.0000000
#>  [7,] 0.4295141 0.4295141 0.97852877  0.576306935 0.8897883 0.8897883
#>  [8,] 0.4295141 0.4295141 0.97852877  0.576306935 0.8897883 0.8897883
#>  [9,] 0.0839088 0.0839088 0.09419942 -0.002649054 0.1120377 0.1120377
#> [10,] 0.5677384 0.5677384 0.83073248  0.542567600 0.9709926 0.9709926
#>             [,7]       [,8]         [,9]      [,10]
#>  [1,] 0.42951406 0.42951406  0.083908803 0.56773844
#>  [2,] 0.42951406 0.42951406  0.083908803 0.56773844
#>  [3,] 0.97852877 0.97852877  0.094199417 0.83073248
#>  [4,] 0.57630693 0.57630693 -0.002649054 0.54256760
#>  [5,] 0.88978831 0.88978831  0.112037706 0.97099261
#>  [6,] 0.88978831 0.88978831  0.112037706 0.97099261
#>  [7,] 1.00000000 1.00000000  0.081835914 0.85857486
#>  [8,] 1.00000000 1.00000000  0.081835914 0.85857486
#>  [9,] 0.08183591 0.08183591  1.000000000 0.06275431
#> [10,] 0.85857486 0.85857486  0.062754312 1.00000000
#> 
#> $rr
#>  [1] 1.006941 1.006941 1.003325 1.005205 1.001454 1.001454 1.004674 1.004674
#>  [9] 1.089820 1.034184
#>