ldsep: Linkage Disequilibrium Shrinkage Estimation for Polyploids

Estimate haplotypic or composite pairwise linkage disequilibrium (LD) in polyploids, using either genotypes or genotype likelihoods. Support is provided to estimate the popular measures of LD: the LD coefficient D, the standardized LD coefficient D', and the Pearson correlation coefficient r. All estimates are returned with corresponding standard errors. These estimates and standard errors can then be used for shrinkage estimation. The main functions are ldfast(), ldest(), mldest(), sldest(), plot.lddf(), format_lddf(), and ldshrink(). Details of the methods are available in Gerard (2021a) doi:10.1111/1755-0998.13349 and Gerard (2021b) doi:10.1038/s41437-021-00462-5 .

See also

Author

Maintainer: David Gerard gerard.1787@gmail.com (ORCID)