Linkage Disequilibrium Shrinkage Estimation for Polyploids

Estimate haplotypic or composite pairwise linkage disequilibrium (LD) in polyploids, using either genotypes or genotype likelihoods. Support is provided to estimate the popular measures of LD: the LD coefficient D, the standardized LD coefficient D', and the Pearson correlation coefficient r. All estimates are returned with corresponding standard errors. These estimates and standard errors can then be used for shrinkage estimation.

Functions

The main functions are:

ldfast()

Fast, moment-based, bias-corrected LD LD estimates from marginal posterior distributions.

ldest()

Estimates pairwise LD.

mldest()

Iteratively apply ldest() across many pairs of SNPs.

sldest()

Iteratively apply ldest() along a sliding window of fixed length.

plot.lddf()

Plot method for the output of mldest() and sldest().

format_lddf()

Format the output of mldest() and sldest() into a matrix.

ldshrink()

Shrink correlation estimates using adaptive shrinkage (Stephens, 2017; Dey and Stephens, 2018).

Citation

If you find the methods in this package useful, please run the following in R for citation information: citation("ldsep")

Author

David Gerard